fix MSVC warnings

+++ core/libs/facesengine/preprocessing/shape-predictor/fullobjectdetection.cpp
@@ -57,7 +57,7 @@ cv::Rect& FullObjectDetection::get_rect()
 
 unsigned long FullObjectDetection::num_parts() const
 {
-    return parts.size();
+    return (unsigned long)parts.size();
 }
 
 const std::vector<float>& FullObjectDetection::part(unsigned long idx) const
@@ -84,10 +84,10 @@ std::vector<cv::Rect> getEyes(const FullObjectDetection& shape)
 
         // initializing
         std::vector<float> firstpoint = shape.part(start);
-        tlx                           = firstpoint[0];
-        brx                           = firstpoint[0];
-        tly                           = firstpoint[1];
-        bry                           = firstpoint[1];
+        tlx                           = (int)firstpoint[0];
+        brx                           = (int)firstpoint[0];
+        tly                           = (int)firstpoint[1];
+        bry                           = (int)firstpoint[1];
 
         for (int i = start ; i <= end ; ++i)
         {
@@ -95,20 +95,20 @@ std::vector<cv::Rect> getEyes(const FullObjectDetection& shape)
 
             if      (x[0] < tlx)
             {
-                tlx = x[0];
+                tlx = (int)x[0];
             }
             else if (x[0] > brx)
             {
-                brx = x[0];
+                brx = (int)x[0];
             }
 
             if      (x[1] < tly)
             {
-                tly = x[1];
+                tly = (int)x[1];
             }
             else if (x[1] > bry)
             {
-                bry = x[1];
+                bry = (int)x[1];
             }
         }
